pyproject.toml
[tool.poetry]
name = "graduate-work"
version = "0.1.0"
description = ""
authors = ["Andrew Zaiko <andrewzaiko@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.25.1"
social-auth-app-django = "^5.0.0"
django-filter = "^22.1"
pydantic = "^1.10.6"
accessify = "^0.3.1"
pytest-django = "^4.5.2"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
